edited creation process, removed jwt validation

+++ src/main/java/org/example/userservice/service/CardService.java
@@ -6,15 +6,17 @@ import org.example.userservice.dto.CardDTO.CardRequestDTO;
 import org.example.userservice.dto.CardDTO.CardResponseDTO;
 import org.example.userservice.exception.BadRequestException;
 import org.example.userservice.exception.EntityNotFoundException;
+import org.example.userservice.exception.ForbiddenException;
 import org.example.userservice.mapper.CardMapper;
 import org.example.userservice.model.Card;
 import org.example.userservice.model.User;
 import org.example.userservice.repository.CardRepository;
 import org.example.userservice.repository.UserRepository;
+import org.example.userservice.security.AuthInfo;
+import org.example.userservice.security.SecurityService;
 import org.springframework.cache.CacheManager;
-import org.springframework.cache.annotation.CacheEvict;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
+
 import java.util.List;
 import java.util.Optional;
 
@@ -25,42 +27,63 @@ public class CardService {
     private final CardRepository cardRepository;
     private final CardMapper cardMapper;
     private final UserRepository userRepository;
+    private final AuthInfo authInfo;
     private CacheManager cacheManager;
+    private SecurityService securityService;
 
-    @PreAuthorize("@securityService.isCardOwnerOrAdmin(#id, authentication)")
     public CardResponseDTO getById(Long id) {
-        return cardRepository
+        Card card = cardRepository
                 .findById(id)
-                .map(cardMapper::toCardResponseTo)
                 .orElseThrow(() -> new EntityNotFoundException("Card with id " + id + " not found"));
+
+        if ( !securityService.isOwner(card.getUser().getId()) ){
+            throw new ForbiddenException("You do not have permission to access this card");
+        }
+
+        return cardMapper.toCardResponseTo(card);
     }
-    @PreAuthorize("hasRole('ROLE_iw.admin')")
     public List<CardResponseDTO> getByIds(List<Long> ids) {
+
         return cardRepository
                 .getByIds(ids)
+                .peek(card -> {
+                    if (!securityService.isOwner(card.getUser().getId())) {
+                        throw new ForbiddenException("You do not have permission to access card " + card.getId());
+                    }
+                })
                 .map(cardMapper::toCardResponseTo)
                 .toList();
     }
 
-    @CacheEvict(value = "userWithCards", key = "#input.userId")
-    @PreAuthorize("@securityService.isOwnerByEmailOrAdmin(#input.userId, authentication)")
     public CardResponseDTO create(CardRequestDTO input) {
-        User user = userRepository.findById(input.getUserId())
-                .orElseThrow(() -> new BadRequestException("User with id " + input.getUserId() + " does not exist"));
+        Long userId = authInfo.getUserId();
+
+        if(userId == null) {
+            throw new BadRequestException("Missing userId in request headers");
+        }
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new BadRequestException("User with id " + userId + " does not exist"));
         Card card = cardMapper.toCard(input);
         card.setUser(user);
         Card savedCard = cardRepository.save(card);
+
+        Optional.ofNullable(cacheManager.getCache("userWithCards"))
+                .ifPresent(cache -> cache.evict(userId));
+
         return cardMapper.toCardResponseTo(savedCard);
     }
 
     @Transactional
-    @PreAuthorize("@securityService.isCardOwnerOrAdmin(#id, authentication)")
     public void delete(Long id) {
 
         var card = cardRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Card with id " + id + " not found"));
 
         Long userId = card.getUser().getId();
+
+        if ( ! ( securityService.isOwner(userId)|| securityService.hasRole("ROLE_iw.admin") ) ){
+            throw new ForbiddenException("You do not have permission to delete this card");
+        }
 
         cardRepository.deleteById(id);
